test : `OrderTable` 테스트 코드 추가 및 리팩토링
