+++ src/test/java/kitchenpos/ordertable/acceptance/step/TableAcceptanceStep.java
@@ -1,4 +1,4 @@
-package kitchenpos.acceptance.step;
+package kitchenpos.ordertable.acceptance.step;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -7,6 +7,7 @@ import io.restassured.mapper.TypeRef;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
+import kitchenpos.acceptance.step.HttpUtil;
 import kitchenpos.ordertable.dto.OrderTableRequest;
 import kitchenpos.ordertable.dto.OrderTableResponse;
